유저 정보 수정/삭제 진행 시 존재하지 않는 유저에 대한 요청 예외처리

+++ IntelliJ/library-app/src/main/java/com/group/libraryapp/controller/user/UserController.java
@@ -64,12 +64,21 @@ public class UserController {
 
     @PutMapping("/user")
     public void updateUser(@RequestBody UserUpdateRequest request) {
+        String checksql = "SELECT * FROM user WHERE id = ?";
+        // jdbcTemplate.query(sql, RowMapper 구현 익명클래스)는 List를 반환한다.
+        if (jdbcTemplate.query(checksql, (rs, rowNum) -> 0, request.getId()).isEmpty()) { // 리스트가 비어있으면
+          throw new IllegalArgumentException();
+        }
         String sql = "UPDATE user SET name = ? WHERE id = ?";
         jdbcTemplate.update(sql, request.getName(), request.getId());
     }
 
     @DeleteMapping("/user")
     public void deleteUser(@RequestParam String name) {
+        String checksql = "SELECT * FROM user WHERE name = ?";
+        if (jdbcTemplate.query(checksql, (rs, rowNum) -> 0, name).isEmpty()) {
+            throw new IllegalArgumentException();
+        }
         String sql = "DELETE FROM user WHERE name = ?";
         jdbcTemplate.update(sql, name);
     }
